chore(deps): 全量依赖升级至最新并统一 minor 级写法
- 跨版本: lbug 0.18→0.20.4（自研，graph-viewer server 联动）、tree-sitter
  0.26→0.27、criterion 0.5→0.8、oxiarc-zstd 0.3→0.4、pest/pest_derive
  2.7→2.9、serial_test 3→4.0、tree-sitter-ocaml 0.25→0.26、
  rust_decimal 1.37→1.43、uuid 1.23→1.26
- 自研四库对齐 base 开发线: trait-kit 0.5.0-rc.5、sdforge/oxcache
  0.5.0-rc.4、inklog 0.3.0-rc.4
- 写法统一 minor 级（"1"→"1.53"、"2"→"2.3"、"3"→"3.27" 等，
  消除全部 major/patch 级写法）
- API 适配: tree-sitter 0.27 Node::kind 改借用节点生命周期
  （complexity.rs Halstead 收集去 &'static 化）；LanguageError 新增
  NotParseable 变体（parse/error.rs 测试补分支）；child_count 改回返回
  u32（daemon.rs count_tree_nodes、rust_extractor、typescript 去多余
  u32 转换）；criterion 0.8 弃用 criterion::black_box（7 个 bench 换
  std::hint::black_box）
- 例外: ort 维持 2.0.0-rc.13（crates.io max_stable_version=null、
  1.16.3 已 yanked，rc 即最新）；自研 rc 写法语法必须三段式
- 验证: 复跑 cargo outdated 直接依赖 0 行待更新；
  cargo test --lib --all-features 4570 passed / 0 failed

+++ src/parse/rust_extractor.rs
@@ -1587,7 +1587,7 @@ fn extern_language(node: Node, source: &str) -> Language {
         }
     }
     // Also check direct string_literal children (older grammar versions).
-    for i in 0..node.child_count() as u32 {
+    for i in 0..node.child_count() {
         if let Some(child) = node.child(i) {
             if child.kind() == "string_literal" {
                 let text = node_text(child, source).unwrap_or("");
@@ -1654,7 +1654,7 @@ fn is_external_pub(node: Node) -> bool {
                 // visibility_modifier's children include the `pub` token,
                 // followed by optional `(`, restriction path, `)`. If any
                 // `(` token is present, there's a restriction.
-                for j in 0..child.child_count() as u32 {
+                for j in 0..child.child_count() {
                     if let Some(token) = child.child(j) {
                         if token.kind() == "(" {
                             return false;
